Move get_payloads to consensus module
