+++ client/src/node.rs
@@ -5,7 +5,6 @@ use std::time::Duration;
 use ethers::prelude::{Address, U256};
 use ethers::types::{Filter, Log, Transaction, TransactionReceipt, H256};
 use eyre::{eyre, Result};
-use futures::future::join_all;
 
 use common::errors::BlockNotFoundError;
 use common::types::{BlockTag, Bytes32};
@@ -116,7 +115,7 @@ impl Node {
             .insert(finalized_payload.block_number, finalized_payload);
 
         let start_slot = self.current_slot.unwrap_or(latest_header.slot - self.history_size as u64);
-        let backfill_payloads = self.get_payloads(start_slot, latest_header.slot).await?;
+        let backfill_payloads = self.consensus.get_payloads(start_slot, latest_header.slot).await.map_err(NodeError::ConsensusPayloadError)?;
         for payload in backfill_payloads {
             self.payloads.insert(payload.block_number, payload);
         }
@@ -136,48 +135,6 @@ impl Node {
         Ok(())
     }
 
-    async fn get_payloads(&self, start_slot: u64, end_slot: u64) -> Result<Vec<ExecutionPayload>, NodeError> {
-        let payloads_fut = (start_slot..end_slot).rev().into_iter().map(|slot| async move {
-            let self_ref = self;
-            let inner_closure = async move {
-                let payload = self_ref
-                    .consensus
-                    .get_block_from_rpc(slot)
-                    .await
-                    .map_err(NodeError::ConsensusPayloadError)?
-                    .body
-                    .execution_payload;
-                Ok::<ExecutionPayload, NodeError>(payload)
-            };
-            inner_closure.await
-        });
-        let mut prev_parent_hash: Option<Bytes32> = None;
-        let mut payloads: Vec<ExecutionPayload> = Vec::new();
-        for result in join_all(payloads_fut).await {
-            if result.is_err() {
-                // If fetching any payload failed we only backfill up to that payload
-                // TODO: Review if it is better to error out of the whole payload update instead;
-                break;
-            }
-            let payload = result?;
-            if let Some(parent_block_hash) = &prev_parent_hash {
-                if &payload.block_hash != parent_block_hash {
-                    // If any of the blocks don't match we abort the whole backfill
-                    // TODO: Review if it is better to still backfill the payloads up to this point
-                    return Err(NodeError::ConsensusPayloadError(
-                        ConsensusError::InvalidHeaderHash(
-                            format!("{:02X?}", parent_block_hash.to_vec()),
-                            format!("{:02X?}", payload.parent_hash.to_vec()),
-                        )
-                        .into(),
-                    ));
-                }
-            }
-            prev_parent_hash = Some(payload.parent_hash.clone());
-            payloads.push(payload);
-        }
-        return Ok(payloads);
-    }
 
     pub async fn call(&self, opts: &CallOpts, block: BlockTag) -> Result<Vec<u8>, NodeError> {
         self.check_blocktag_age(&block)?;
