+++ consensus/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 eyre = "0.6.8"
+futures = "0.3.23"
 serde = { version = "1.0.143", features = ["derive"] }
 serde_json = "1.0.85"
 hex = "0.4.3"
